Integrate @cvasilak comments
